FMWK-788-new-limiter - removed unused casting

+++ pkg/service/restoreexecutor/xdr.go
@@ -57,7 +57,7 @@ func makeXdrRestoreConfig(restoreRequest *model.RestoreRequest,
 	}
 
 	config.RecordsPerSecond = util.ValueOrZero(restoreRequest.Policy.Tps)
-	config.Bandwidth = int64(util.ValueOrZero(restoreRequest.Policy.Bandwidth))
+	config.Bandwidth = util.ValueOrZero(restoreRequest.Policy.Bandwidth)
 	config.Parallel = restoreRequest.Policy.GetParallelOrDefault()
 	config.MaxAsyncBatches = restoreRequest.Policy.GetMaxAsyncBatchesOrDefault()
 	config.BatchSize = restoreRequest.Policy.GetBatchSizeOrDefault()
